Split enormous admin.lua into smaller modules
Module `cartridge/admin.lua` is too big. With this patch it's split into
several smaller `cartridge/lua-api/*` modules.

Public API is unchanged.  Old `admin.lua` is preserved for even better
compatibility.
