+++ cartridge/webui/gql-stat.lua
@@ -1,8 +1,8 @@
 #!/usr/bin/env tarantool
 
 local pool = require('cartridge.pool')
-local admin = require('cartridge.admin')
 local gql_types = require('cartridge.graphql.types')
+local lua_api_stat = require('cartridge.lua-api.stat')
 
 local statistics_schema = {
     kind = gql_types.object({
@@ -84,7 +84,7 @@ local statistics_schema = {
         if cache.servers_stat ~= nil then
             return cache.servers_stat[root.uri]
         elseif cache.disable_stat_optimization then
-            return admin.get_stat(root.uri), nil
+            return lua_api_stat.get_stat(root.uri), nil
         end
 
         local uri_list = {}
